pyproject.toml
[project]
name = "backend-chatbot-sikolog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==3.2.0",
    "fastapi>=0.135.3",
    "langchain-core>=1.3.0",
    "langchain-groq>=1.1.2",
    "langchain-redis>=0.2.5",
    "langgraph>=1.1.9",
    "langgraph-checkpoint-redis>=0.4.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.13.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.26",
    "redis>=7.4.0",
    "supabase>=2.28.3",
    "uvicorn>=0.44.0",
]
